requirements.txt
aiohttp==3.7.4.post0
anyio==3.3.4
asgiref==3.5.0
async-timeout==3.0.1
attrs==21.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
cheroot==8.5.2
CherryPy==18.6.1
construct==2.10.67
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
discord==1.7.3
discord.py==1.7.3
Django==4.0.2
django-filter==21.1
django-phonenumber-field==5.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
ed25519==1.5
h11==0.12.0
httpcore==0.13.7
httpx==0.20.0
idna==3.3
importlib-metadata==4.11.1
iniconfig==1.1.1
itypes==1.2.0
jaraco.classes==3.2.1
jaraco.collections==3.4.0
jaraco.functools==3.3.0
jaraco.text==3.5.1
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.1.0
more-itertools==8.10.0
multidict==6.0.2
openapi-codec==1.3.2
packaging==21.2
pluggy==1.0.0
portend==3.0.0
py==1.10.0
pycparser==2.20
PyJWT==2.3.0
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==6.2.5
pytz==2021.3
requests==2.26.0
rfc3986==1.5.0
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
solana==0.18.1
sqlparse==0.4.2
tempora==4.1.2
toml==0.10.2
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
websockets==10.0
yarl==1.7.2
zc.lockfile==2.0
zipp==3.7.0
